pyproject.toml
[tool.poetry]
name = "chatdrug"
version = "0.1.0"
description = ""
authors = ["xiaojin <xiaojinwhu10@gmail.com>"]
packages = [
    { include = "chatdrug" },
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<=3.12"
torch = "^2.0.1"
nvidia-cublas-cu11 = "^11.11.3.6"
nvidia-cuda-cupti-cu11 = "11.7.101"
nvidia-cuda-nvrtc-cu11 = "^11.7"
nvidia-cuda-runtime-cu11 = "^11.7"
nvidia-cudnn-cu11 = "^8.9.1.23"
nvidia-cufft-cu11 = "^10.9.0.58"
nvidia-curand-cu11 = "^10.3.0.86"
nvidia-cusolver-cu11 = "^11.4.1.48"
nvidia-cusparse-cu11 = "^11.7.5.86"
nvidia-nccl-cu11 = "^2.16.5"
nvidia-nvtx-cu11 = "^11.7"
langchain = "^0.0.186"
pandas = "^2.0.2"
docarray = "^0.33.0"
tiktoken = "^0.4.0"
sentence-transformers = "^2.2.2"
pinecone-client = "^2.2.2"
chroma = "^0.2.0"
chromadb = "^0.3.26"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ipykernel = "^6.23.1"
dvc = "^2.58.1"
openai = "^0.27.7"
arxiv = "^1.4.7"
pypdf = "^3.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
